Minimal patch to work with base p2pool by forrestv
This should get this UI into a minimal working state for the base p2pool by forrestv (https://github.com/forrestv/p2pool)
There are UI elements that need to be cleaned up when the data simply isn't present, that will come next.
This patch at least makes it so it doesn't crash out of the gate.

References Issue #1